Reworked capi cached handles management.
Handles are persistent references to managed-memory objects (i.e.
"Ruby" objects). Objects that have handles are kept alive either
by the handle having a greater than 0 reference count, or the object
being reachable from another object.

Further, there are essentially two types of handles: 1) those that
have cached data, and 2) those that do not. Handles can have cached
data when they support direct memory access. For example Strings and
Arrays accessed by RSTRING and RARRAY. This is provided purely for
compatibility with poorly written MRI C extensions.

Each NativeMethod call (i.e. a call to a C function in a C extension)
tracks which handles were requested during that call. The handle's
reference count is incremented when the handle is requested and
decremented when that native method call returns. Refernces to these
handles are kept in an STL Set class.

There is also a global list of handles maintained by the garbage
collector. This enables the GC to keep the handles is the list up-to-date
when the object it references is garbage collected.

When a regular handle is referenced in a way that creates cached data,
the handle is moved from the global list to a cached-data list. Both
lists are processed by the various GC stages. However, only the cached
data list is processed at native method call boundaries. When control
returns to Ruby-land from C-land, any cached data must be flushed back
to the Ruby objects so it is visible to Ruby code that manipulates them.
When control goes from Ruby-land to C-land, the cached data must be
updated from the Ruby objects so that C functions that operate on the
cached data will have the correct data.