Changed benchmark monitor to use fork/exec.
If the bench runner process exceeds the time limit, that process
will be killed and the benchmark status recorded as Timeout. If
the bench runner process exits with a non-zero status, that status
is recorded. Otherwise the status of the bench is 'success'.

If a benchmark uses system() to run another process and that
benchmark exceeds the time limit, the subprocess will not be killed.
This would potentially pollute later benchmark results. This needs
a solution.