Revert changes to Array::concat primitive.
This reverts commits b93a407 and 947f356.

The code was instrumented to record sizes for 0, 1, N, Z, where
N was inside the capacity checking branch added by the reverted
commits, Z was falling through to the bottom, and 0, 1 were at
the top (i.e. in the position before the code was modified by
the reverted commits).

The following programs were run with the instrumentation:

  rdoc benchmark
  mspec ci -B full pre-compiled
  mspec ci -B full compiling
  rails -h
  gem install rdoc remote

The usage data:

rdoc benchmark

0:   51796
1:  644811
N:       3
Z:   30063

mspec ci pre-compiled

0:   62333
1:   66381
N:       0
Z:    1372

mspec ci compiling

0:   59579
1:  543210
N:       0
Z:    1373

rails -h

0:    5046
1:   57907
N:       8
Z:   24616

gem install

0:    5524
1:   56089
N:       1
Z:    6252

While it is true that some of the cases where 1 element was passed
may have exceeded the size of the tuple and necessitated resizing
(the inverse of fitting and not needing resizing), that logic is
well handled by Array::set.

Further, this code highlights important cases, namely how many items
are passed, not how much space there is in the tuple. When writing a
JIT version or porting to Ruby code, this communicates the important
cases.

The space-checking code may make some logical sense, but the cost of
checking is much higher than checking if size == 1 and delegating.

Finally, the usage data is fairly consistent across several programs.
But even if the usage data changes, the core cases and the amount of
work to handle each is better accounted for by the original code.