This reverts the rbc generation option, see below.
Recommendations: Since *not* writing .rbc files is the option (ie not the default
behavior), add a :no_rbc option to .unified_load that defaults to false. Adjust
the conditionals appropriately. Then track all paths that call unified_load and
ensure that your option is being passed appropriately. The current Compiler arch
complicates this a bit.