Local variables the melbourne way.
This is substantially still a WIP. Previously, the AST was created
from the sexp top-down (i.e. each child node could depend on the
parent node existing when the child is created). The AST is now
created from the MRI parse tree bottom-up (i.e. all child nodes
are created before the parent node).

For local variables, this inversion creates some confusion. A local
variable in code is actually a complex web of abstractions: a scope
that contains storage for the local variables, a local variable as
a member of that scope with certain attributes, a reference to a
local variable at the point of access or assignment, and possibly a
nested scope.

This will be documented in detail once it is complete. For now, more
work needs to be done getting method definitions, array literals,
splat and svalue nodes, and class/module bodies working correctly
before more lasgn specs can be fixed. Also note that dynamic locals
(i.e. eval locals) are not yet implemented.