Resurrect ConcatArgs. All array compiler specs pass.
Like a nauseatingly sweet cake, the compiler currently has
3 thick layers of obfuscating logic: 1) the sexp rewriting,
2) the sexp consumption and args methods, 3) the bytecode
generation. The compiler is a bad marriage between a sexp
processor and an object-oriented compiler. All this is built
on the often confusing foundation of MRI's parse tree.

I mean, seriously, rewrite argscat into splats that are wrapped
in arrays so other nodes can take them apart again? Seriously?

So I am the divorce judge and this union is being destroyed.
Sorry, sexps, you get nothing. Dismissed. Goodbye. Objects,
you get stripped of your twisty internal logic, to the extent
reasonable. How? By creating more classes as appropriate that
each do their own simple thing.

Again, this is a WIP, but the general principle is that if a
thing needs a particular form, it should be a particular thing,
not some murky shape-shifter.