Fixed svalue, splat bytecode.
Some compiler specs fail, but they are broken and need to be
fixed. As a consequence of wrapping splat nodes in array nodes
in the rewriter, literals of the form [*[something]] were
causing the bytecode sequence to be emitted:

  make_array 0     # yup, create an empty array
  push something   # ie push each elt of the literal array
  make_array size  # size of the inner literal array
  send :+          # put all the inner elts in the empty array

This is totally unnecessary since

  [] + [something] == [something]

The specs will be updated once the compiler changes are made
active.