Refactored iter args and some masgn fixups.
One of the goals of refactoring the compiler front end is to use
logic while processing the parse tree to create a clear AST form
where each node uses minimal logic to generate bytecode. In other
words, represent the form of the resulting instructions in the
structure of the AST rather than in conditionals during bytecode
output.

In the case of masgn and iter args (which is simply a special case
of masgn), this could be done (as it was with the initial iter args
approach). However, this approach would be much better applied at
the parse tree level (which essentially means fixing the parser, and
that ain't gonna happen in the near term). Reforming the parse tree
into a different AST form required a lot of substitution and results
in a very similar structure. The tradeoff is a more complex
MAsgn#bytecode method.  However, the logic is much simpler than the
existing bytecode generation that has to deal with the confusion
introduced by the Sexp rewriting.

So, while I'm not totally happy with this approach, it is fairly
clean, much simpler, and should perform better than rewriting
every masgn.