Removed push_unique_literal from compiler specs.
This is an implementation detail, similar to push 1 emitting meta_push_1. At
most, the difference between push_literal and push_unique_literal will be the
index into the literals tuple. The important pieces of data are what is being
pushed and the fact it is put in the literals tuple.