Hack around bug caused by inlining
We must prevent the JIT from inlining __script__, since it's always
different. We're going around the deopt framework by accessing the
method_table directly.

But even if we did, we currently don't deopt running methods properly.
So the new script would be put in the method_table, then #as_script
would be deopt'd for the NEXT run, but we need it deopt'd in the current
running to run the proper script.

Rather than worrying about the deopt strategy, hack around it by using
__send__, which the JIT won't see through to inline __script__.