Fix missing block errors caused by block inlining
When we inline methods and their blocks, we don't create the
BlockEnvironment for them, because we usually don't need them. But one
place we do need them is if a guard is hit and we are deopt'ing to the
interpreter. In that case, we need for all BlockEnviroments in the
current inlining scope to be created, because any of them might be
needed by the method as it drops back to the interpreter (and thus no
longer runs a previously inlined block)