Replace set_const insn with bootstrap methods.
Rubinius generally handles bootstrapping issues by defining a
method with restricted functionality that is present while the
core library is loading, then defining a full-featured version
in the core library (often in kernel/delta). This change removes
the compiler special-casing for setting constants by defining
the necessary methods using the VM's bootstrap method facility,
which basically attaches a primitive method directly to a method
table.

For simplicity and clarity, this mechanism should be used uniformly.