Added bootstrap methods to handle #attach_method.
This removes the #kernel? special-casing in the compiler for
defining singleton methods. In other words:

Q. How are singleton methods defined on a metaclass?
A. By calling Metaclass#attach_method.

The bootstrap Metaclass#attach_method is redefined in
kernel/delta/metaclass.rb as normal.

This does not change the use of Rubinius.attach_method
for __class_init__, __module_init__, and __metaclass_init__.
But those are handled consistently in kernel and userland.