Added ruby_version_is guard. See below.
The previous ruby_bug guard is questionably correct. It is more accurate
to say that MRI behavior changed than that the previous behavior was a
bug that was fixed. The change in #1648 is more an API change than it is
something like 1 + 2 == 4.

Since all alternative implementations are expected to pass all ruby_bug
guarded specs, the ruby_version_is guard also makes more sense, since
the different behavior was added in 1.8.8dev.