Use better way of pop'ing exception handlers in the JIT
It was possible to loose a pop_unwind because it was detected as dead
code, which meant that handlers weren't being unregistered properly.

This caused infinite loops checking handlers. This makes it so that JIT
entry to codegen the landing pad of an exception handler causes the
handlers list to be pop'd, so it can't reenter itself.