Simplify masgn output in *a = b cases.
In the case of *a = b (and all variations like *@a = b), the right-hand side
is always wrapped in an array, so there is no need to do any tests.

Furthermore, the current compiler output is demonstrably wrong. For example,

  ruby -e 'class A; def to_ary; return 1, 2; end; end; *a = A.new; p a'

Output should be '[#<A:0x___]' whereas current rbx calls #to_ary.