Improve young GC and add more GC stats
This adds an eden area to the young GC. This means that post collection,
the young GC always has a fixed number of bytes available to new
allocations. This fixes a pathalogical case where collection frees
almost nothing, and therefore there is almost no space for new objects.
This causes the young GC to collect very, very often and copying a lot
of data back and forth without getting anything done.