Add header inflation and faster capi access
Refactored access to bits of the header word and added the ability to
inflated that word into a pointer. The pointer points to an
InflatedHeader object that the GC manages non-moving, mark/sweep wise.

Access to a capi::Handle* for an object is therefore accessible directly
via the InflatedHeader.

Future code will likely put a lock into the InflatedHeader for object
locking.