CAUTION: Speed Zone Ahead.
This enables CompilerNG as the default compiler. There is presently
an issue either with Melbourne and the C-API or with the compiler
causing abnormal load on the GC mature generation. This means that
full compiling CI runs, for instance, will be slower. We are actively
working on addressing this issue and ultimately the C-API, GC, compiler
or all three will be better as a result.

Note that this commit hacks in CompilerNG as the default. This will
get prettier as more things are refactored. Promise.