Add magic comments and Array Zen
The parser now communicates any magic comments it sees to the compiler.
Currently, the compiler enables any transform available using the text
of the comment as the transform name.

The current transform using this is Array Zen. It's a list comprehension
transform. Currently, limited, but should be expanded to handle a number
of more forms.