Make error msg more descriptive when runtime is not found.
Rubinius is built with a hard-coded path for where it should locate the core
(kernel) library files and the standard library. If the directory is moved,
rbx will be unable to find the files. However, you can set RBX_RUNTIME to
the location of the kernel files and RBX_LIB to the location of the standard
library and rbx should run.