Fixed running IRB from install.
This should be considered a provsional fix. See the comment
in the install task. One problem with this approach is that
the assumption is the .rb file will have an older date than
the .rbc file. Another problem is that it simply masks very
bad practice, like looking specifically for a something.rb
file instead of relying on the require machinery to do its
job. Rubinius will be susceptible to these sorts of bugs.