Changed require paths for compiler to avoid RUBYOPT complications.
The precompile scheme for installing Rubinius would use -Ilib with MRI. If the
user had RUBYOPT=rubygems set, MRI would require rubygems which would end up
trying to load the Rubinius lib/etc.rb.

This approach essentially hardcodes the paths relative to the actual files.
This is desirable given that the compiler needs to be loaded in various
Ruby implementations (probably MRI but not necessarily so) and we should
ensure that only these particulay files are loaded.