Fix ugly JIT crash bug
the method #build_message in Test::Unit exposed this bug, though I'm
pretty sure I've seen it around before.

Basically, if a method took optional args AND a splat, then the code
would blindly use the passed arg count as the loop counter, even if
there weren't that many arguments! This caused CallFrame corruption,
because part of the CallFrame was override with args there was no room
for.

The simple fix is to clamp the loop counter with the total number of
args we know the method takes staticly.