Moved spec/frozen to spec/ruby. See below.
To simplify working with specs in Rubinius, changes to specs should be
committed in separate commits to the appropriate files in ./spec/ruby.
These changes will be regularly pushed to the RubySpec project. Changes
to RubySpec are regularly sync'd with the files in ./spec/ruby.

The rake tasks for cloning RubySpec inside the Rubinius project have
been removed.