Changed compiler specs related to "#{a}". See below.
The string_dup instruction in these is superfluous, but it is
necessary in dstr (literally "#{a}") as shown in commits a241091f
and 34654ea5. Here's a code example that demonstrates:

  # vastly simplified insane code courtesy of mkmf.rb
  def checking_message(target, place)
    [" in"].inject("#{target}") do |m, p|
      m << p
    end
  end

  def foo(var)
    checking_message var, "foo"
  end

  a = "some_string"     # the local 'a' now points to a dup of "some_string"
  foo a                 # 'a' is passed to #foo, which passes it to #checking_message
  p a                   # #checking_message modifies 'a' to be "same_string in"

One possibility would be to refactor #bytecode in DynamicString
and use a special case for one-element string interpolation. As
the comment in 34654ea5 states, I think these optimizations are
dubious.