Add stack locals, fix current exception logic
This adds the ability to have stack locals. These are local variables,
indexed purely by number, that don't follow ruby semantics. They can not
be closed over, and are allocated by adding extra slots at the end of
the operand stack.

The only use of these stack locals is to save and restore the current
exception when entering and exiting a rescue/ensure.