Big memory usage cleanup
This fixes a memory leak caused by VMMethod. To fix the leak, to major
changes were added.

1) A CodeManager was introduced, which integrates with the Immix GC.
This CodeManager acts like a mark/sweep for VMMethod C++ objects,
allowing them to be deleted when the CompiledMethod they're used by goes
out of scope.

2) VMMethod no longer holds a TypedRoot<> to the CompiledMethod it's
for. This root was keeping all CompiledMethods alive, so no VMMethods
were cleaned up, and thusly #1 is useless without this changes. This
touched a lot of files because the JIT used VMMethod::original a bunch.