Add String#to_inum support for soft bases. Fixes #163.
MRI, strangely enough, treats String#oct as a soft request for a base of
8. If a base identifier is found on the string, the 8 is ignored and
whatever the base identifier says is used.

"0b1010".oct # => 10

Swell. So String#to_inum now supports the same protocol rb_str_to_inum
in MRI does, which is that if the base is negative, then it's a soft
base and can be overridden if there is a base identifier found.
Otherwise the requested base is flipped to positive and used.