Tag a, b = x failures pending decision on new insn.
We presently emit the same bytecode for 'a, b = x' and 'a, b = *x' but
these are not semantically the same. Since the syntactic context is
lost by the time the cast_array insn executes, I think we need a new
insn, but I'm waiting to see what Evan thinks.