Fix missing write_barrier calls in the JIT
The JIT builds a set of RuntimeData objects which contain the
CompiledMethod, Symbol, and Module an inlined method is for. The set of
these objects is associated with the CompiledMethod that the methods
were inlined into.

Because the RuntimeData contain references to these managed objects, the
write barrier must be run to associate the rooted CompiledMethod with
the information about inlined methods. Otherwise, the RuntimeData is not
updated when CompiledMethod's for inlined methods are moved by the GC.