Speed up Array#hash
This inlines a bunch of recursion detection related code for the sake of
improving performance. It's not well refactored therefore but it's about
7x faster than the previous code.

To you, the reader of this commit message who is wondering why Evan has
done this unfactoring of code, please benchmark your replacement code if
you'd like to improve on this. Don't simply say "this is so unDRY!" and
rewrite it to be 7x slower.

Thanks!
  - MGMT