Fix File.readlink. Fixes #256.
FFI used to be able to manipuate a String objects insides, thats how
this method expected to work. We removed that functionality at some time
in the past because it's too hard to make work properly (and it's not
compatible with ruby-ffi).

The proper way is to use an explicit MemoryPointer and read the data
back out of it, as is done here.