Teach onig to treat invalid utf-8 sequences are raw bytes
This allows onig to mimick the 1.8 (and improve on it) by validating
utf-8 prefix bytes and their trail bytes. If the trail bytes are invalid
for the prefix byte (ie, they don't start with 10 in their bit pattern),
then the prefix and the trail bytes are seen instead as raw bytes. This
thusly doesn't allow a utf-8 prefix byte to "hide" any ASCII characters
after it.