Fix building with gcc 4.4.1.
According to http://gcc.gnu.org/bugzilla/show_bug.cgi?id=25509,
gcc 4.5 added a flag -Wno-unused-result to turn off this warning.
This flag appears to work fine with 4.4.1 on Ubuntu 9.10. However,
the flag is not supported in earlier versions of gcc. The attempt
to work around this warning with (void)foo(blah); does not work
for glibc functions as discussed in the ticket above. So, we can
either sniff the gcc version and enable this flag if supported,
stop using gcc altogether, or just use the result (however
vacuously). This approach uses the result.