Track and verify threads stopping, allow running code when alone.
A ManagedThread now has a run_state that is tracked to tell where what
it is doing. We verify that all are stopped except the requester on
stop_the_world. Additionally, checkpointing is ignored when alone, which
allows a thread to stop everyone else and then run some ruby code.