Raise a TypeError if attempting to inherit from a metaclass.
This primitive, set_superclass, is a bit complicated. We should strive
to always give the user the best error info that we can. In this case,
the primitive can fail if the superclass is not a Class or if it is a
MetaClass, which is a Class that is not visible to the user (and presently,
not testable in Ruby).

If we merely return Primitives::failure for all cases, the two possible
errors are confounded. We could also raise exceptions from the C++ code
for both cases, but this requires more complexity to set up the appropriate
error message if the superclass is not a Class. Further, there is really
no valid Ruby "fallback code" in the case of this primitive.

Presently, we have no well-documented guidelines for these situations.