Specs for CSV#readlines for 1.9. Closes #431.
The original patch for this ticket had a lot of problems. Code like

  x.each { |z| z.class.should == String or NilClass}

makes no sense. Moreover, testing the results by checking the class
with x.class is not the way specs should be written.

This CSV library has a number of significant changes between the version
bundled with 1.8 and the version in 1.9. The specs need some helpers to
create CSV files but may need different versions of the helpers for 1.8
and 1.9.