Flush/update CAPI handles better. Fixes #455.
Previous to this, we were flushing all cached handles whenever a capi
function exited. We had a bug where we should have been updating all
capi handles when a capi function was entered. We now do 2 sets of
logic. The first is a faster, more simple logic of updating/flushing
handles seen by the current NativeFrame. This includes non-local handles
passed to R*(). This should deal with all reasonable extensions fine.

We also introduce -Xcapi.global_flush which when on will cause all
cached handles to be flushed/update on capi call boundaries.