Emit error message when attempting to run .rbc file.
The .rbc files are a caching mechanism and no code should depend on
their existence. The .rbc files are low-tech and convenient but we
may be able to gain significant speed and memory load advantages by
using a completely different mechanism. If code starts depending on
.rbc files, then that code would break. There is simply no good
reason for running .rbc files directly.