Backport add --cxx to configure from multiverse (6596ed7). Closes #542.
From original commit:

Most versions of GCC are smart enough to understand when we are compiling/linking
C++ code. However, the GCC 4.5 in MingW is apparently not one of the smart ones
so the tr1/hash detection code fails on Windows. On the other hand, C++ changed
the C prototype for ANYARGS functions, so e.g. BigDecimal C ext will not compile
properly with g++. So, we provide options for both in configure.