More fixes for configure under Windows.
All subprocess commands in configure are run with cmd.exe except
for running rakelib/config.guess which requires no environment setup
but does rely on sh.

Our general approach to building on Windows now requires the MinGW
tools. But some of these commands require PATH or other environment
settings. To simplify things, we expect that the environment for
cmd.exe be set up to build. The build should also work under MSYS
bash or cygwin bash or something else, but we aren't going to try
to maintain environments for all those. Since cmd.exe is on all
Windows installs, that's the lowest common denominator.