Abstract native threads more consistently.
There are basically four levels of abstraction over threads currently. The
list below goes from least to most abstract:

1. the underlying native thread implementation; presently, pthreads.
2. thread::Thread and friends in vm/util/thread.hpp. This abstracts
   the OS threads.
3. rubinius::NativeThread in vm/native_thread.[ch]pp inherits from
   thread::Thread and provides abstraction at the VM level below Ruby
   semantics.
4. rubinius::Thread utilizes rubinius::NativeThread to provide Ruby
   Thread semantics.

The thread::Thread and related abstractions provide some OO niceness but
essentially are just the functionality of pthread. On Windows, we could
abstract at the level of thread::Thread or use a compatible pthread
implementation. The advantage of utilizing (or writing) a pthread-compatible
library is consistency and less likelihood of threading peculiarities
getting masked with different implementations of thread::Thread.

In this scheme, pthread should not appear anywhere above thread::Thread.