Revert "Explicitly set the class, it could be a subclass of Proc"
We need more specs to show what needs to be fixed and what will break.

This commit breaks the following code in Rake:

    define_module_function :system, &Kernel.method(:system)

When the Method#to_proc is called and creates a Proc::Method object,
the BlockEnvironment (Proc#block) is nil.

This reverts commit 463eb57ba40d198778a43da3a5ed00bca50e0db7.

Revert "Remove tag for now passing Proc spec"

This reverts commit 77ee296cf8dacf6bf83e84b8ada705dcb70bd981.