More better RSTRING caching.
Introduce a macro that C extensions can define to disable updating
and flushing the C-API RString cache for a Ruby string. The default
is to update and flush the cache.

There are two primary scenarios involved in the C-API:

  1. Ruby -> C -> Ruby ...
  2. Ruby -> C -> C++ (ie. vm/builtin/*) -> Ruby

In general, refences to Ruby objects in C code are mediated by a
separate C++ structure called a Handle. When objects pass the boundary
from Ruby into C, handles are created for them.

For some objects, notably String, the MRI C extensions are commonly
coded to access the underlying C structure directly. Since Rubinius
cannot allow this, we provide a secondary structure that mimics the
MRI RStruct structure. We term this a cache. There are two operations
that must be performed on the cache: 1) updating the cache with the
current contents of the Ruby object; and 2) sync'ing the Ruby object
with the current contents of the cache (ie RString). By default, the
update operation occurs when going from Ruby to C, and the sync
operation occurs when returning from C to Ruby.

However, some of the Rubinius C-API is implemented by calling directly
into the C++ code that implements the primitive operations for objects
like String. In these cases, the contents of the cache must be manually
sync'd to the Ruby object before calling the C++ code and the cache
must be manually updated after the C++ call before returning to the
C extension.