Use mspec :launch instead of :flags so Agent doesn't start in sub-specs.
There are specs that unconditionally kill the exe and that does
not allow the exe to perform any cleanup. This results in numerous
garbage rbx Agent connection files. It would be nice to have the
Agent running in all spec processes in case one hangs, but since
rbx doesn't get a chance to do cleanup when :KILL'd, there's no
good way to prevent littering these files. So Agent is only started
for the main process running the specs.