Fix Thread::wakeup and random thread failures
Turns out that linux requires a thread to be detached or joined. It can
not be left orphaned in a potentional joinable state as linux will not
clean it up automatically (it's easy to justify why this is).

Thus, we set all threads as detached and use our own join logic as
before.