Concurrency 101: know your locked region
We were, rather I was, stupidly manipulating internal Lock bookkeeping
outside of the locking itself. This created a race condition where one
thread would unlock a Lock and another thread would lock it BEFORE we
reset locking_thread to 0 in the original thread. This confused the
system and cause a deadlock.