Improve thread debugging, fixed 2 thread bugs
Bug 1: If contention was detected, we would unlock the header using CAS
(e1), then call into ObjectMemory to inflate the header (e2) and install a
native mutex. The code assumed e2 was directly followed by e1, but the
bug arose when between e1 and e2, another thread saw the unlockd header
and CAS'd the header with it's thread id (e3).

This confused e2 because it was then trying to inflated a header that
was locked by an unknown thread. This caused the header and inflated
header to enter an unknown state, which in turn caused an unlock
failure.

Bug 2: If while waiting for a contented lock to become free GC was
performed, the ObjectMemory::mutex_ would remain locked and cause a
deadlock. This was because we used SYNC(state) in contend_for_lock and
then called GCIndependent with OM's mutex locked. The events would play
out such that WorldState::become_dependent would be waiting for
waiting_to_run_ since the other thread was waiting for the OM lock to
become free.

The fix here was to introduce another lock that is only used when
registering and resolving contention.