Add hand mantained types.ll and offset gen code
Previously, we'd use LLVM to read the code base and output an .ll file
for each of the structs we'd want. Then we'd turn that .ll file into a
.cpp file that would recreate the types for LLVM to use. Using llvm-g++
for this has proved difficult because of structal typing collapses. So
instead we're going to hand maintain a types.ll with the types and field
names of the things want to access in the JIT.

Additionally, the comments with the field names are read auto
offset_specific.hpp is output, making it easier to access a field in a
struct by it's name (this will replace the hand maintained offset.hpp)