Remove unnecessary memory barriers
We don't need a memory barrier right after atomic::compare_and_swap
since that already is a memory barrier in itself. We also need the
barrier for unlocking it.

Worse case that could happen when unlocking is that due to out of
order reads we see the unlock one lock iteration later. Taking this
hit occasionally is cheaper than always taking a memory barrier hit
on each unlock.