Hack around yyerror arity and %parse-param.
MRI defines %parse-param and also defines:

  yyerror(msg)  parser_yyerror(<parse-param>, msg)

The generated parser calls parser_yyerror(parser, msg) instead
of yyerror(parser, msg). But how to get bison to generate like
this is beyond me. For grammar19.y, bison generates calls to
yyerror(parser_state, msg), which is invalid in the face of the
definition of yyerror(msg). This work around uses a different
macro (yy_error) in the syntax rules.