Fix EBADF error by being more careful with fd's and IO objects
Because IOs are always finalized, it's really important that an IO
object doesn't point to an fd that has been closed outside of the IO
object itself. Previously, the IO.setup to run setsockopt would point
the IO to an fd that we later might manually close, creating an EBADF
error race.