Implementation of insert-order hashes.
Turns the hash object into a linked list on top of the hash buckets. Each hash entry contains a next and prev pointer, and the hash itself has a pointer to the first and last. Iteration happens by walking this list rather than walking the buckets as previously.

Some different tradeoffs could be made, like not including a previous pointer on the items and turning delete into an O(n) operation instead of an O(1) operation, but this doesn't really seem worth it to me. It does seem unfortunate to add two extra pointers to each item, though.