Switch to memory friendly Node tree for profiler
Node replaces Edge and forms a tree start from a root Node. The data for
which method a Node is for is held in a seperate Method object can
therefore be shared by all Nodes for the same Method, which fixes the
memory usage problem.

Also, this commit introduces -Xprofiler.json for dumping the profiling
output as JSON for offline processing.