Use object's #respond_to? in Type.try_convert.
The Type predicates are used to protect critical parts of the Rubinius kernel
from careless or stupid user code. In this case, since the (ridiculous) pattern
of overriding both #respond_to? and #method_missing is somewhat engrained, we
permit the potentially overridden #respond_to? to be called. If the user code
behaves inconsistently, that's the user's problem.