Some parser state cleanup.
Starting with 1.9, MRI added a parser state struct like we had
in Melbourne and accesses the previous global variables through
macros. We had macros for some of the variables but the access
was ad hoc and inconsistent. This cleanup attempts to make the
access consistent. However, there are some variables like 'enc'
that are still accessed via directly parser_state->.