Fixed dynamic regexp specs.
These changes should also be made for 1.8. The argument could
be made that this should not be supported, but I would disagree.
Since #to_s can be changed on nil, not following these semantics
would lead to incompatible behavior in an edge case. Likewise,
there is no real drawback to supporting this as it would probably
be rarely used.