VM Instructions Documentation Localization Draft
Here's the draft proposal of how we could localize VM instruction
documentation.

Why?
====

Mostly because the auto-generated instructions.markdown is all English and I'd
like it (particularly descriptions and notes) to be translated into my
language ('ru') too. The second reason is my love to Ruby hacking but that's
only the second ;-)

What?
=====

We could generate additional files with the names like
`_site/instructions_#{lang}.markdown` containing Descriptions and Notes of a
corresponding language. Then we would be able to `{% include %}` the generated
file into VM/Instructions article. As the result, the whole article wold be in
the same language except of what should be in English (code, name, parameters,
etc.)

How?
====

I've modified `rakelib/vm.rake` and `rakelib/instruction_parser.rb` files so
we could use the following syntax:

 1. In  `vm/instructions.def`:

	# [Description]
	#   Pushes the value of the integer literal onto the stack.
	# [Description:ru]
	#   Помещает значение целочисленного литерала в стек.
	# [See Also]
	#   meta_push_0
	#   meta_push_1
	#   meta_push_2
	#   meta_push_neg_1
	# [Notes]
	#   Certain common cases (i.e. -1, 0, 1, and 2) are optimised to avoid
	#   the decoding of the argument.
	# [Notes:ru]
	#   Некоторые частые случаи применения (-1, 0, 1, 2) оптимизированы на
	#   предмет избегания декодирования аргумента.

    As you can see, we can specify Descriptions and Notes language as
    `:#{lang}` in the section name. If language label is absent, the script
    assumes it's English.
 2. In `rakelib/vm.rake`:

    * As usual:

          file "web/_includes/instructions.markdown" => insn_deps do |t|
	    generate_instruction_file iparser, :generate_documentation, t.name
	  end

    * And with the language specification:

	  file "web/_includes/instructions_ru.markdown" => insn_deps do |t|
	    generate_instruction_file iparser, :generate_documentation, t.name, 'ru'
	  end

    In the latter case the file will be generated containing only Descriptions
    and Notes in the specific language, as it's written in the source.

This solution is backward-compatible with the old, language-agnostic syntax.
In this case the language is assumed to be English (`lang = 'en'`) and the
resulting markdown will contain English instruction descriptions and notes.

Result?
=======

You can see the result in _includes/instructions_ru.markdown and
web/doc/ru/virtual-machine/instructions/index.html of this commit.