Modify the result string in place when using to_sub_replacement
This greatly improves performance if a lot of small parts
are replaced in a string using sub / gsub. Before to_sub_replacement
allocated a string object in each iteration, which could cause
a lot of GC pressure in large strings with a lot of replacements.

The performance of the benchmark added for single character replacements
in the previous comment increases by 50% with this change.