Make sure that high InflatedHeader allocations trigger a GC
This fixes issue #1080. The problem was that the script provided
in this issue has such a GC pressure profile that running young
generation GC's seemed to be enough. Due to using the CAPI it was
allocating a lot of InflatedHeaders, but those allocations weren't
taking into consideration when determining to do which GC.

This fixes this issue by tracking the number of chunks that are
allocated for InflatedHeaders. The current threshold is pretty
arbitrary, right know it will trigger a mature GC after 16 * 1024
InflatedHeaders are allocated.