Fixed issue with ThreadGroup.
Previously, the call the #prune in eg #list would check if the weakref and object
were still alive. However, it's possible for GC to run after the call to #prune
and before the @threads iteration, resulting in an exception being thrown when
a collected object is referenced. This occurred in the Travis CI framework tests.

Since #prune is not generally useful as a utility method, I removed it completely
and inlined it into #add. Also, there are no users of #list_all and the method
does not exist in MRI, so I removed it.