Fixed hanging Thread spec.
The problem with this spec is that ScratchPad.record is being used to
signal that the thread is actually running. The signalling is done by
setting the value recorded to :running. But if an exception is raised,
the value :running is then overwritten with the exception object. The
other thread, busy waiting until the value is set to :running then
busy waits forever. This race is avoided by using a separate mechanism
to signal between threads.

In general, ScratchPad should only be used for computed values or for
collecting state values (eg by instrumenting control flow). For other
cases, like signaling here, we have to make custom mechanisms.

The thread specs also need to be completely rewritten.