Check whether any handlers need to run when read() was interrupted
This was causing a hang in cases where the parent rbx process was
sent a signal, because the signal was never handled and the rbx
process therefore never stopped. The fix is based on how we handle
this in various other places where a system call can be interrupted.

This also fixes that there was a case where an exception based on
an errno was raised inside a block where the GIL wasn't acquired.