Added CI tags for block arg edge case.
Essentially, the issue here is that despite 1.8 using masgn for block args
there is a significant difference in semantics. This can be seen as follows:

  def r(a) yield(*a) end
  r([[]]) { |a, b| p a, b }

In MRI, this would print

  []
  nil

In rbx, this prints

  nil
  nil

The issue is, we are using

  cast_for_multi_block_arg
  shift_array
  set_local 0
  ...

The problem is that masgn has different behavior here. In the case of

  a, b = []

a and b will be assigned 'nil', just like in rbx for the block above.