Make Object::to_s aware of Singleton classes and use it in the profiler
By making Object::to_s aware of Singleton classes, it generates a nicer
output for this case. Instead of generating #<Class(Class):0xXX> it now
generates #<Class(String):0xXX> for a Singleton class for a String.

This commit also changes the profiler to use this method instead of
calling to_s on the instance. This fixes the case where a \000 byte
is present in for example a String object.

In general I think this change also is better since it shows on which
class a method is defined in the profiler output instead of the instance
in case of a method on a Singleton class.

Fixes #1118