Make sure to setup the list of threads in SharedState properly
When we fork when multiple threads are running, the threads list
was copied over to the fork. When the time came the fork wanted to
exit, this resulted in the infamous "Tried to stop but threads still
running" error message, since the threads from the original process
where still in the thread list.

This commit clears the thread list and sets it up properly again so
that we can always exit nice and clean.

Fixes #1000