Make sure to restart the SignalHandler after execvp() failed
Otherwise we lack signal handler in a process that tried to
do an execvp() and failed. We mark the SignalHandler thread
as not deleting the thread when it exits, since we might need
to restart it. This isn't a risk for a leak, since we never
deallocate the SignalHandler in a process anyway.